+++ lavis/models/sevila_models/sevila.py
@@ -68,7 +68,7 @@ class SeViLA(Blip2Base):
         # freeze T5
         for name, param in self.t5_model.named_parameters():
             param.requires_grad = False
-            param.data = param.data.float() 
+            param.data = param.data.bfloat16() 
 
         # Q-Former for Answerer
         self.Qformer, self.query_tokens = self.init_Qformer(
@@ -158,7 +158,7 @@ class SeViLA(Blip2Base):
                 answer = samples['qa_output']
                 ans_idx = [self.ANS_MAP[a[-1]] for a in answer]
 
-                with torch.cuda.amp.autocast(dtype=torch.float32):
+                with torch.cuda.amp.autocast(dtype=torch.bfloat16):
                     # Frame Prefix
                     frame_prefix = self.t5_tokenizer(
                         self.frame_prefix, padding="longest", add_special_tokens=False,
@@ -231,7 +231,7 @@ class SeViLA(Blip2Base):
                 encoder_attention_mask=image_atts, return_dict=True) # bt, n, c
             inputs_t5_loc = self.t5_proj_loc(query_output_loc.last_hidden_state) # bt, n, c
             atts_t5_loc = torch.ones(inputs_t5_loc.size()[:-1], dtype=torch.long).to(image.device)
-            with torch.cuda.amp.autocast(dtype=torch.float32):
+            with torch.cuda.amp.autocast(dtype=torch.bfloat16):
                 frame_prefix = self.t5_tokenizer(
                     self.frame_prefix, padding="longest", add_special_tokens=False,
                     truncation=True, max_length=self.max_txt_len, return_tensors="pt").to(image.device) 
@@ -280,7 +280,7 @@ class SeViLA(Blip2Base):
                 inputs_t5_loc = self.t5_proj_loc(query_output_loc.last_hidden_state)
 
                 atts_t5_loc = torch.ones(inputs_t5_loc.size()[:-1], dtype=torch.long).to(image.device)
-                with torch.cuda.amp.autocast(dtype=torch.float32):
+                with torch.cuda.amp.autocast(dtype=torch.bfloat16):
 
                     frame_prefix = self.t5_tokenizer(
                         self.frame_prefix, padding="longest", add_special_tokens=False,
@@ -384,7 +384,7 @@ class SeViLA(Blip2Base):
             del subtitle_tokens_qa, subtitle_ids_qa
             torch.cuda.empty_cache()
 
-            with torch.cuda.amp.autocast(dtype=torch.float32):        
+            with torch.cuda.amp.autocast(dtype=torch.bfloat16):        
                 vid_prefix = self.t5_tokenizer(
                     self.vid_prefix, padding="longest", add_special_tokens=False,
                     truncation=True, max_length=self.max_txt_len, return_tensors="pt",).to(image.device) # 
@@ -441,7 +441,7 @@ class SeViLA(Blip2Base):
             text_input_qa = samples['qa_input'] 
             answer = samples['qa_output'] 
 
-            with torch.cuda.amp.autocast(dtype=torch.float32):
+            with torch.cuda.amp.autocast(dtype=torch.bfloat16):
                     # Frame Prefix
                 if 'qa_vid' not in self.task:
                     atts_t5_qa = torch.ones(inputs_t5_qa.size()[:-1], dtype=torch.long).to(image.device) 
@@ -551,7 +551,7 @@ class SeViLA(Blip2Base):
                 encoder_attention_mask=image_atts, return_dict=True)
             inputs_t5_qa = self.t5_proj(query_output_qa.last_hidden_state)
             
-            with torch.cuda.amp.autocast(dtype=torch.float32):
+            with torch.cuda.amp.autocast(dtype=torch.bfloat16):
                     # Frame Prefix
                 if 'vid' not in self.task: 
                     atts_t5_qa = torch.ones(inputs_t5_qa.size()[:-1], dtype=torch.long).to(image.device) 
@@ -633,7 +633,7 @@ class SeViLA(Blip2Base):
             inputs_t5_loc = self.t5_proj_loc(query_output_loc.last_hidden_state)
 
             atts_t5_loc = torch.ones(inputs_t5_loc.size()[:-1], dtype=torch.long).to(image.device)
-            with torch.cuda.amp.autocast(dtype=torch.float32):
+            with torch.cuda.amp.autocast(dtype=torch.bfloat16):
 
                 frame_prefix = self.t5_tokenizer(
                     self.frame_prefix, padding="longest", add_special_tokens=False,
@@ -884,7 +884,7 @@ class SeViLA(Blip2Base):
         inputs_t5_loc = self.t5_proj_loc(query_output_loc.last_hidden_state)
 
         atts_t5_loc = torch.ones(inputs_t5_loc.size()[:-1], dtype=torch.long).to(image.device)
-        with torch.cuda.amp.autocast(dtype=torch.float32):
+        with torch.cuda.amp.autocast(dtype=torch.bfloat16):
 
             frame_prefix = self.t5_tokenizer(
                 self.frame_prefix, padding="longest", add_special_tokens=False,
@@ -1061,7 +1061,7 @@ class SeViLA(Blip2Base):
         encoder_atts = torch.cat([atts_t5, input_tokens.attention_mask], dim=1)
 
         device_type = "cuda" if "cuda" in str(self.device) else "cpu"
-        with torch.amp.autocast(device_type=device_type, dtype=torch.float32):
+        with torch.amp.autocast(device_type=device_type, dtype=torch.bfloat16):
             inputs_embeds = self.t5_model.encoder.embed_tokens(input_tokens.input_ids)
             inputs_embeds = torch.cat([inputs_t5, inputs_embeds], dim=1)
